Make registry maker and remove stale code
gd-extension-type-registry should be the final representation of gd
extension interface header. I think this might be buggy because typedefs
can have const in them but that is outside the scope of this commit.

I also realized that it only makes sense to define const for variables
and pointers. We don't have variables in the header so it's only
pointers. The const is applied to type but the modifier doesn't bear any
meaning if this type doesn't have additional context (in this case
pointer is pointing).